RT-16 auto increase bib and change to bib to int

+++ lib/data/dto/participant_dto.dart
@@ -14,7 +14,7 @@ class ParticipantDto {
   static Participant fromJson( Map<String, dynamic> json, String id) {
     return Participant(
       id: id,
-      bib: json['bib'],
+      bib: json['bib'] as int,
       firstName: json['firstName'],
       lastName: json['lastName'],
       age: json['age'],
